🔨 refactor: my page 코드 분리
