+++ src/app/(user)/(service)/my-page/liked/page.tsx
@@ -1,23 +1,15 @@
 "use client";
 
-import FeedbackList from "@/src/features/feedback/feedback-list/components/feedback-list";
 import LikedFeedbackList from "@/src/features/my-page/my-page-feedback/components/my-page-feedback-list";
 import MyPlacePagination from "@/src/features/my-place/my-place-common/components/my-place-pagination";
-import { ArrowLeftIcon, ArrowRightIcon } from "@/src/components/utils/icons";
-import client from "@/src/lib/api/client";
-import { IFeedbackListItem } from "@/src/types/feedback/feedbackList";
-import { keepPreviousData, useQuery } from "@tanstack/react-query";
+import { ArrowLeftIcon } from "@/src/components/utils/icons";
 import { useSession } from "next-auth/react";
 import Link from "next/link";
 import { useState } from "react";
-
-type MyPlageResponse = {
-  searchResults: IFeedbackListItem[];
-  totalPages: number;
-};
+import { useLikedFeedbackQuery } from "@/src/features/my-page/my-page-feedback/queries/liked-feedback";
 
 export default function Page() {
-  const { data: session, status } = useSession();
+  const { data: session } = useSession();
   const [page, setPage] = useState(0);
   const size = 15;
 
@@ -26,43 +18,13 @@ export default function Page() {
     isLoading,
     isError,
     error,
-  } = useQuery<MyPlageResponse>({
-    queryKey: ["myFeedbacks", session?.user?.memberId, page],
-    queryFn: async () => {
-      if (!session?.user?.memberId)
-        return {
-          searchResults: [],
-          totalPages: 1,
-        };
-
-      const response = await client.get(
-        `/api/members/${session?.user.memberId}/feedbacks/liked`,
-        {
-          params: { page, size },
-        },
-      );
-
-      if (response.data.status === "성공") {
-        return {
-          searchResults: response.data.data.searchResults,
-          totalPages: response.data.data.totalPages,
-        };
-      }
-      throw new Error("최근 경로 불러오기 실패");
-    },
-    placeholderData: keepPreviousData,
-    enabled: !!session?.user?.memberId,
-  });
-
-  const { searchResults, totalPages } = feedbackList;
-
-  console.log(totalPages);
+  } = useLikedFeedbackQuery(session?.user?.memberId, page, size);
 
   if (isLoading) return <div>로딩 중...</div>;
-
   if (isError && error instanceof Error)
     return <div>에러가 발생했습니다: {error.message}</div>;
-  console.log(session?.user?.memberId);
+
+  const { searchResults, totalPages } = feedbackList;
 
   return (
     <div className="flex w-full justify-center">
