+++ src/app/(user)/(service)/my-page/settings/page.tsx
@@ -1,50 +1,20 @@
 "use client";
 
-import { Switch } from "@/src/components/ui/switch";
 import { ArrowLeftIcon } from "@/src/components/utils/icons";
-import { signOut, useSession } from "next-auth/react";
+import { useSession } from "next-auth/react";
 import Link from "next/link";
-import { useMutation } from "@tanstack/react-query";
-import client from "@/src/lib/api/client";
-import Swal from "sweetalert2";
 import LogoutButton from "@/src/features/auth/auth-common/components/logout-btn";
-import { DropdownThemeToggle } from "@/src/features/display-mode/display-mode-common/components/display-mode-toggle";
+import { DropdownThemeToggle } from "@/src/features/my-page/my-page-setting/components/my-page-display-mode-toggle";
+import {
+  handleDeleteUser,
+  useDeleteUser,
+} from "@/src/features/my-page/my-page-setting/actions/delete-user";
 
 export default function Page() {
   const session = useSession();
   const userId = session.data?.user.memberId;
 
-  const deleteUserMutation = useMutation({
-    mutationFn: async (userId: number) => {
-      if (!session?.data?.user?.memberId) return;
-      await client.delete(`/api/members/${session.data?.user.memberId}`);
-    },
-    onSuccess: () => {
-      signOut({ redirectTo: "/login" });
-    },
-    onError: (error) => {
-      console.error(error);
-      alert("탈퇴에 실패했습니다. 다시 시도해주세요.");
-    },
-  });
-
-  const handleDeleteUser = () => {
-    Swal.fire({
-      title: "정말 탈퇴하시겠습니까?",
-      text: "더 이상 즐겨찾기와 피드백 작성을 이용할 수 없게 됩니다.",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#FF6156",
-      cancelButtonColor: "#64748B",
-      confirmButtonText: "탈퇴",
-      cancelButtonText: "취소",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        if (!userId) return;
-        deleteUserMutation.mutate(userId);
-      }
-    });
-  };
+  const deleteUserMutation = useDeleteUser();
 
   return (
     <div className="flex w-full justify-center">
@@ -74,7 +44,9 @@ export default function Page() {
         <LogoutButton />
         <div className="flex justify-center">
           <button
-            onClick={handleDeleteUser}
+            onClick={() =>
+              handleDeleteUser(userId, (id) => deleteUserMutation.mutate(id))
+            }
             className="theme-my-page-withdraw mt-[222px] text-xs font-medium hover:text-red"
           >
             회원 탈퇴
