+++ src/features/my-page/my-page-direction/components/my-page-bookmark-card.tsx
@@ -1,0 +1,32 @@
+interface Props {
+  name: string;
+  address?: string;
+  onClick: () => void;
+}
+
+export default function DestinationCard({ name, address, onClick }: Props) {
+  return (
+    <div onClick={onClick} className="flex gap-2">
+      <div className="flex aspect-square w-[40px] items-center justify-center rounded-full bg-teal">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="#FFFFFF"
+          viewBox="0 0 24 24"
+          className="h-[24px] w-[24px]"
+        >
+          <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M8.161 2.58a1.875 1.875 0 0 1 1.678 0l4.993 2.498c.106.052.23.052.336 0l3.869-1.935A1.875 1.875 0 0 1 21.75 4.82v12.485c0 .71-.401 1.36-1.037 1.677l-4.875 2.437a1.875 1.875 0 0 1-1.676 0l-4.994-2.497a.375.375 0 0 0-.336 0l-3.868 1.935A1.875 1.875 0 0 1 2.25 19.18V6.695c0-.71.401-1.36 1.036-1.677l4.875-2.437ZM9 6a.75.75 0 0 1 .75.75V15a.75.75 0 0 1-1.5 0V6.75A.75.75 0 0 1 9 6Zm6.75 3a.75.75 0 0 0-1.5 0v8.25a.75.75 0 0 0 1.5 0V9Z"
+          />
+        </svg>
+      </div>
+      <div>
+        <p className="theme-my-profile-location-name font-bold">{name}</p>
+        <p className="theme-my-profile-location-address text-xs font-medium">
+          {address}
+        </p>
+      </div>
+    </div>
+  );
+}
