+++ src/features/my-page/my-page-direction/index.tsx
@@ -1,10 +1,7 @@
 "use client";
 
-import client from "@/src/lib/api/client";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { useSession } from "next-auth/react";
-import { useEffect } from "react";
 import RecentDestinationsItem from "@/src/features/my-page/my-page-direction/components/my-page-direction-item";
+import useRecentPaths from "@/src/features/my-page/my-page-direction/queries/recent-paths";
 
 export interface Destination {
   recentPathId: number;
@@ -17,46 +14,9 @@ export interface Destination {
 }
 
 export default function RecentDestinations() {
-  const { data: session, status } = useSession();
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    console.log(session);
-    if (status === "authenticated") {
-      queryClient.invalidateQueries({
-        queryKey: ["recentPaths", session?.user?.memberId],
-      });
-    }
-  }, [session?.user?.token]);
-
-  const {
-    data: destinations = [],
-    isLoading,
-    isError,
-    error,
-  } = useQuery<Destination[]>({
-    queryKey: ["recentPaths", session?.user?.memberId],
-    queryFn: async () => {
-      if (!session?.user?.memberId) return [];
-      const response = await client.get(
-        `/api/members/${session.user.memberId}/recent-path`,
-        {
-          headers: {
-            Authorization: `Bearer ${session.user.token}`, // 최신 토큰 반영
-          },
-        },
-      );
-      const apiData = response.data;
-      if (apiData.status === "성공") {
-        return apiData.data as Destination[];
-      }
-      throw new Error("최근 경로 불러오기 실패");
-    },
-    enabled: !!session?.user?.memberId && status === "authenticated",
-  });
+  const { destinations, isLoading, isError, error } = useRecentPaths();
 
   if (isLoading) return <div>로딩 중...</div>;
-
   if (isError && error instanceof Error)
     return <div>에러가 발생했습니다: {error.message}</div>;
